requirement.txt
asgiref==3.7.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.47.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysqlclient==2.2.1
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.21.12
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.3
